Write regression spec for correct new name in transfer message
This test was kind of brutal to write because I tried for an isolated
test that doesn't rely on any previous state. Hopefully, subsequent
messages will be easier if they follow some patterns here.

I extracted some helper methods out of `before()` callbacks for
re-usability and we should be able to keep extracting more of these
until we have better test data.

Also, I brought in (what looks to be a reasonably easy) mocking library
so as to start testing the interface of message sending rather than
integrating it as part of the controller test.