Change "closed" to "archived" in operations.js
Searching the codebase made me notice that this is the only place the
"closed" URL argument is used. In the clients controller it appears to
be a synonym of "archived".

I like "archived" more since it implies less finality, so to standardize
I am going to convert this link to use that.