Add initial chef server configuration
This adds some basic server configuration as specified in the Chef
recipe DSL, which is a good starting point for getting a
production-hardened instance of clientcomm running.

Chef works by allowing you to specify a set of "resources" in recipes
with certain properties (e.g. file ownership, chmod) and then an action
to perform on that resource to bring the machine into convergence (e.g.
"create", "delete").

The Chef TODO list is still roughly:

- [*] Non-root user that the app will run as
- [*] Installing a fixed version of nodejs
- [*] Downloading a copy of the code
- [ ] Nginx reverse proxy in front
- [ ] Systemd init script

(although there is no real reason this commit could not be merged. It
works.) And then we will have to figure out if we want to use an
application deployment framework like capistrano/a nodejs equivalent, or
just roll our own, which wouldn't be too difficult for the short term.