Add terraform configuration for S3 and credentials file
One dependency of Clientcomm is an S3 bucket to share with Twilio for
call recordings. To do this properly we will create an IAM user with the
minimum permissions necessary to read/write from a bucket.

These credentials can be passed in via environment variables to the
clientcomm process. Included is a credentials.js template that extracts
all the values from the environment.