Fix terraform credential file location
First, we have to log into the instance with the "ubuntu" user. This
will require the user to have added the private key to their SSH agent.

Secondly, the bucket interpolation should be the "bucket" attribute
instead of the nonexistent "name" attribute.

Thirdly, terraform gives no way to move the file with "sudo" so sticking
it in the home directory will be a good place for the chef cookbook to
expect.