Provision ELB, multiple app servers with terraform
This commit introduces a "count" parameter in a couple places to
provision multiple resources in different availability zones.
Specifically, we want to create multiple web servers in different AZ's,
so we can set `count = 2` and then configure each instance's subnet to
be different. In turn, we need to provision multiple subnets, one in
each AZ.

To handle the multiple subnet concepts, I am carving up the VPC IP
space:

  10.1.x.x -> web servers
  10.2.x.x -> database servers (TODO in a subsequent commit)

The ELB will be the entry point and be able to distribute traffic to
whichever servers are available at the time.