Add SSL/TLS configuration instructions
Unfortunately, there is some manual provisioning necessary for this to
work. But thankfully it is free and pretty easy to set up. Instructions
are in the README.md.

Once the `TF_VAR_aws_ssl_certificate_arn` variable is set, everything
should just work by letting the ELB terminate SSL for us. That's not the
best place to terminate SSL, but it is certainly the easiest because it
prevents us from needing to distribute private keys to each app server.