Automatically restart clientcomm service on failure
If the service crashes due to a network connectivity problem, memory
leak, or whathaveyou, we can have systemd restart it automatically.

This will wait 10 seconds before restarting, which is something I've
learned through previous experience will help us not restart the service
too quickly (the default is 100ms) so as to avoid the
StartLimitInterval and keep restarting the service indefinitely until it
comes back up.