Combine receiving DNS records
Since the receiving DNS records are both for the same "name"
("multnomah.clientcomm.org"), they should both be given as elements of
the "records" array rather than being specified as two different
resources. This will prevent the behavior of terraform of always
assuming that updates are required for that record set.