Add test for Messages.sendMessage via sms
There was previously a test for email, but not for SMS. It is somewhat
difficult to stub the twilio module require, so this commit creates a
module in app/lib to be stubbed in tests.

Over time, we can opt-in other twilio-based methods into this module as
we add tests for them.