sms-service module for Twilio API interaction
I'm wrapping all Twilio API interactions in a new sms-service class.
This'll make it easier to mock for tests. There was a previous attempt
at this in twClient.js but I started fresh with a new approach based on
the analytics-service module.

Tests are failing, but most of the heavy lifting has been done.